refactor(ui): split weather/calendar API, centralize keys
Extract weatherRegion and migrationCalendar; wire queryKeys on Overview, WeatherCard, MigrationCalendar, DatabaseMaintenanceCard. Refs #343

Made-with: Cursor
