+++ app/ui/src/api/api.tsx
@@ -6,7 +6,6 @@ import {
   Species,
 } from '../types';
 import axios from 'axios';
-import { formatLocalTime } from '../util';
 import type { TimeOfDay } from '../utils/timeUtils';
 import type { components } from '../generated/openapi-types';
 import { BASE_API_URL, BASE_URL } from './client';
@@ -69,60 +68,17 @@ export {
 } from './timeline';
 export type { UnknownDetection } from './timeline';
 
-export const fetchWeather = async () => {
-  const response = await axios.get(`${BASE_API_URL}/weather`);
-  return response.data;
-};
-
-/** Sunrise, sunset, dawn, dusk for date at configured location. date: YYYY-MM-DD. Returns ISO strings (UTC). */
-export const fetchSunTimes = async (
-  date: string,
-): Promise<{
-  dawn?: string;
-  sunrise?: string;
-  noon?: string;
-  sunset?: string;
-  dusk?: string;
-} | null> => {
-  try {
-    const res = await axios.get(`${BASE_API_URL}/sun-times`, {
-      params: { date },
-    });
-    const d = res.data;
-    if (!d || typeof d !== 'object' || !('sunrise' in d)) return null;
-    return d;
-  } catch {
-    return null;
-  }
-};
-
-/** Format ISO UTC string to local HH:MM */
-export const formatSunTimeLocal = (iso: string): string => {
-  try {
-    return formatLocalTime(iso);
-  } catch {
-    return '--:--';
-  }
-};
-
-/** Region comparison with eBird. Returns null if API key not configured or error. */
-export const fetchRegionComparison = async (): Promise<{
-  regionCode: string;
-  userCount: number;
-  regionTopCount: number;
-  matchCount: number;
-  matchedSpecies: string[];
-  regionTop: string[];
-} | null> => {
-  try {
-    const res = await axios.get(`${BASE_API_URL}/region-comparison`);
-    const d = res.data;
-    if (!d || typeof d !== 'object' || !('regionCode' in d)) return null;
-    return d;
-  } catch {
-    return null;
-  }
-};
+export {
+  fetchRegionComparison,
+  fetchSunTimes,
+  fetchWeather,
+  formatSunTimeLocal,
+} from './weatherRegion';
+export type {
+  MigrationCalendarData,
+  MigrationCalendarParams,
+} from './migrationCalendar';
+export { fetchMigrationCalendar } from './migrationCalendar';
 
 export {
   deleteVideo,
@@ -1565,31 +1521,6 @@ export const fetchSpectrogramRegenerationStatus =
     );
     return response.data;
   };
-
-export interface MigrationCalendarData {
-  species: Array<{
-    id: number | null;
-    name: string;
-    image_url: string | null;
-    monthly_counts: number[];
-    total: number;
-  }>;
-  month_labels: string[];
-  catalog?: 'observed' | 'dataset' | 'full_eu';
-}
-
-export const fetchMigrationCalendar = async (params?: {
-  start_year?: number;
-  end_year?: number;
-  start_date?: string;
-  end_date?: string;
-  catalog?: 'observed' | 'dataset' | 'full_eu' | 'active' | 'full';
-}): Promise<MigrationCalendarData> => {
-  const response = await axios.get(`${BASE_API_URL}/migration-calendar`, {
-    params: params || {},
-  });
-  return response.data;
-};
 
 export const fetchOverviewData = async (
   date: string,
